fix: BootWindow uint->int cast, PldmEncoder missing constants
- Aspeed_eSPI_BootWindow.cs line 128: fix uint->int implicit cast error
- PldmEncoder.cs: add CmdGetSensorReading, SensorDataSize constants,
  SensorOpState constants, PdrNumericSensor (byte), WriteReal32LE,
  RangeFieldFormat constants, WriteLE64 — fixes pre-existing build errors

+++ src/Emulator/Peripherals/Peripherals/Miscellaneous/Aspeed_eSPI_BootWindow.cs
@@ -125,7 +125,7 @@ namespace Antmicro.Renode.Peripherals.Miscellaneous
             }
 
             uint totalSize = ReadHeaderField(HDR_TOTAL_SIZE);
-            uint crc = ComputeCrc32(memory, HeaderSize, (int)Math.Min(totalSize, WindowSize - HeaderSize));
+            uint crc = ComputeCrc32(memory, (int)HeaderSize, (int)Math.Min(totalSize, (uint)(WindowSize - (int)HeaderSize)));
             WriteHeaderField(HDR_CRC32, crc);
 
             state = BootWindowState.Ready;
